% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitglms.R
\name{fitglms}
\alias{fitglms}
\title{Fit GLMs}
\usage{
fitglms(zz, xx, xx_ppd, BB, nphi, nsill)
}
\arguments{
\item{zz}{An array containing fine resolution simulated Poisson data.}

\item{xx}{A list containing fine resolution simulated (true) covariate data.}

\item{xx_ppd}{A list containing downscaled covariate data.}

\item{BB}{Number of bootstrapped replicates.}

\item{nphi}{Number for nphi.}

\item{nsill}{Number for nsill.}
}
\value{
A list of length = 2.
}
\description{
Fit GLMs
}
